Make calculator work with all number types

+++ Calculator/calc/calc.go
@@ -1,10 +1,10 @@
 package calc
 
-func Add(num1, num2 int) int {
+func Add(num1, num2 float64) float64 {
 	return num1 + num2
 }
 
-func Subtract(num1, num2 int) int {
+func Subtract(num1, num2 float64) float64 {
 	return num1 - num2
 }
